+++ lib/pages/loginpage.dart
@@ -1,4 +1,3 @@
-import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import '../utility/router.dart' as route;
@@ -196,7 +195,6 @@ class LoginPageState extends State<LoginPage> {
           password: password,
         );
         _userInfo.put('uid', userCredential.user!.uid);
-        isAdmin();
         navigateToHomePage();
       } on FirebaseAuthException catch (e) {
         if (e.code == 'user-not-found') {
@@ -214,17 +212,5 @@ class LoginPageState extends State<LoginPage> {
 
   void navigateToRegisterPage() {
     Navigator.pushReplacementNamed(context, route.registerpage);
-  }
-
-  void isAdmin() {
-    FirebaseFirestore.instance
-        .collection('Users')
-        .doc(_userInfo.get('uid'))
-        .get()
-        .then((DocumentSnapshot documentSnapshot) {
-      if (documentSnapshot.exists) {
-        _userInfo.put('isAdmin', documentSnapshot.get('isAdmin'));
-      }
-    });
   }
 }
